bid endpoint
