+++ packages/auctioneer/package.json
@@ -6,6 +6,7 @@
   "license": "MIT",
   "scripts": {
     "dev": "yarn build --watch --preserveWatchOutput",
+    "start": "node dist/index.js",
     "lint": "eslint ./src --ext .ts --env node",
     "test": "nyc ts-mocha --check-leaks --exit --timeout 120000 'test/**/*.spec.ts'",
     "clean": "rimraf ./dist ./tsconfig.tsBuildInfo",
@@ -44,7 +45,9 @@
   },
   "gitHead": "937a7cde93e6ac1e151c6374f48adf83d3fa4ec6",
   "dependencies": {
+    "@connext/utils": "^7.3.17",
     "fastify": "^3.27.1",
+    "fastify-plugin": "^3.0.1",
     "pino": "^7.8.0",
     "tsc": "^2.0.4"
   }
