+++ packages/auctioneer/src/index.ts
@@ -1,27 +1,11 @@
-import { fastify, FastifyInstance } from 'fastify';
-import { Server } from 'http';
-import pino from 'pino';
+import { fastifyStart } from "./createAuctioneer";
+import bidRoute from './handlers/bid';
 
-const REDIS_URL = process.env.REDIS_URL || 'http://localhost:6379';
-const LISTEN_PORT = process.env.PORT || 1234;
-const LOG_LEVEL = process.env.loglevel || 'debug';
+console.log(`Creating Auctioneer`)
 
-const pino_logger = pino({ level: LOG_LEVEL });
-
-async function fastifyStart(): Promise<FastifyInstance>{
-  const server = fastify({ logger: pino_logger});
-  try {
-    await server.listen(LISTEN_PORT);
-    server.log.info({},
-      `Auctioneer Listening @ ${LISTEN_PORT}`
-    );
-    return server;
-  }
-  catch (e) {
-    server.log.error(e);
-  }
-
-}
 export async function makeAuctioneer() {
   const serverInstance = await fastifyStart();
+
 }
+
+makeAuctioneer();
